Extract conversation action type into its own file
