+++ Wire-iOS/Sources/UserInterface/GroupDetails/ConversationActions/ConversationActionController.swift
@@ -87,84 +87,3 @@
     }
     
 }
-
-extension ZMConversation {
-    enum Action {
-        case delete
-        case leave
-        case silence(isSilenced: Bool)
-        case archive(isArchived: Bool)
-        case cancelRequest
-        case block(isBlocked: Bool)
-    }
-    
-    var actions: [Action] {
-        switch conversationType {
-        case .connection: return availableOneToOneActions()
-        case .oneOnOne: return availableOneToOneActions()
-        default: return availableGroupActions()
-        }
-    }
-    
-    private func availableOneToOneActions() -> [Action] {
-        precondition(conversationType == .oneOnOne)
-        var actions = [Action]()
-        if nil == team, let connectedUser = connectedUser {
-            actions.append(.block(isBlocked: connectedUser.isBlocked))
-        }
-        actions.append(contentsOf: availableStandardActions())
-        actions.append(.delete)
-        return actions
-    }
-    
-    private func availablePendingActions() -> [Action] {
-        precondition(conversationType == .connection)
-        return [.archive(isArchived: isArchived), .cancelRequest]
-    }
-    
-    private func availableGroupActions() -> [Action] {
-        var actions = availableStandardActions() + [.delete]
-        if activeParticipants.contains(ZMUser.selfUser()) {
-            actions.append(.leave)
-        }
-        return actions
-    }
-    
-    private func availableStandardActions() -> [Action] {
-        var actions = [Action]()
-        if !isReadOnly {
-            actions.append(.silence(isSilenced: isSilenced))
-        }
-        actions.append(.archive(isArchived: isArchived))
-        return actions
-    }
-}
-
-extension ZMConversation.Action {
-    
-    fileprivate var style: UIAlertActionStyle {
-        switch self {
-        case .delete, .leave: return .destructive
-        default: return .default
-        }
-    }
-    
-    fileprivate var title: String {
-        return localizationKey.localized
-    }
-    
-    private var localizationKey: String {
-        switch self {
-        case .delete: return "meta.menu.delete"
-        case .leave: return "meta.menu.leave"
-        case .silence(isSilenced: let muted): return "meta.menu.silence.\(muted ? "unmute" : "mute")"
-        case .archive(isArchived: let archived): return "meta.menu.\(archived ? "unarchive" : "archive")"
-        case .cancelRequest: return "meta.menu.cancel_connection_request"
-        case .block(isBlocked: let blocked): return blocked ? "profile.unblock_button_title" : "profile.block_dialog.button_block"
-        }
-    }
-    
-    func alertAction(handler: @escaping () -> Void) -> UIAlertAction {
-        return .init(title: title, style: style) { _ in handler() }
-    }
-}
